readme and refactor setup
