+++ waterMark/README.txt
@@ -1,0 +1,10 @@
+ ___           _        _ _
+|_ _|_ __  ___| |_ __ _| | |
+ | || '_ \/ __| __/ _` | | |
+ | || | | \__ \ || (_| | | |
+|___|_| |_|___/\__\__,_|_|_|
+
+Ensure waterMark.py and the setUp directory are in the same folder.
+cd setUp
+Windows run admin powershell python setUp.py 
+Linux	run sudo python3 setUp.py
